fixed randomness, create new rule bases and pass to the clients, other fixes
Fixed randomness for repeatability
parse_rule_base returns a dict with the parameters, index of terms, score and class. Also index are sorted and low is always 0, etc.
a new function to create a rule base from the union of the others
moved model initialization from global to local
function to evaluate the rule base on the clients